Refactor: update test names to reflect v1.0+ architecture
- Merge test_pointcompare.py + test_multimodelcompare.py → test_comparercollection.py
- Merge test_simple_compare.py → test_match.py
- Rename test_trackcompare.py → test_comparer_track.py

All 114 tests preserved. Names updated from pre-1.0 (fmskill) architecture
where PointComparer/TrackComparer were separate classes.
